Move connect Reduce from MenuDrawer to MenuDrawercontainer.
